% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{breakbins}
\alias{breakbins}
\title{Take a filegroup with arbitrary bin length and break it into SC size sections.}
\usage{
breakbins(data, binlen)
}
\arguments{
\item{data}{A dataframe of type FG}

\item{binlen}{length in seconds of bin- standard are 300 (LOW) 225 (REG) and 90 (SHI). Other values will be marked as 'custom'}
}
\value{
FG dataset
}
\description{
Using old (pre-NAS) effort data, convert to FG format.
}
