% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_DML_SQL.R
\name{data_delete}
\alias{data_delete}
\title{Delete records from table.}
\usage{
data_delete(data, table)
}
\arguments{
\item{data}{The data table object in R session}

\item{table}{The destination data table name in the database}
}
\value{
not sure lol
}
\description{
Delete records from a named table using the id of the data.
}
