% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckg_scripts.R
\name{data_insert}
\alias{data_insert}
\title{Insert data into the database}
\usage{
data_insert(data, table)
}
\arguments{
\item{data}{The data table object in R session}

\item{table}{The destination data table name in the database}
}
\value{
confirmation or error.
}
\description{
Insert data into a named database table.
}
