% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_DML_SQL.R
\name{data_modify}
\alias{data_modify}
\title{Modify an existing database table.}
\usage{
data_modify(data, table)
}
\arguments{
\item{data}{The data table object in R session}

\item{table}{The destination data table name in the database}
}
\value{
not sure lol
}
\description{
Modify data into a named database table. This method uses keys to match to existing data, which it deletes and replaces upon matches
}
