% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_DML_SQL.R
\name{data_pull}
\alias{data_pull}
\title{Pull data from query}
\usage{
data_pull(query)
}
\arguments{
\item{query}{The full query}
}
\value{
a dataframe
}
\description{
Pull data from the database using an SQL query (currently SQLite syntax)
}
