% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_DML_SQL.R
\name{FG_insert}
\alias{FG_insert}
\title{Insert a filegroup into the database}
\usage{
FG_insert(FG, Name, SelectionMethod, Description)
}
\arguments{
\item{FG}{the filegroup object (specification:)}

\item{Name}{the name of the filegroup}

\item{SelectionMethod}{the SelectionMethod of the filegroups (ex: random, semi-random, procedural, hand)}

\item{Description}{a short text description of the filegroup.}
}
\value{
the filegroup dataframe object
}
\description{
Insert a filegroup into the database by providing the filegroup table and some table metadata
}
