% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_DML_SQL.R
\name{FG_pull}
\alias{FG_pull}
\title{Filegroup Pull}
\usage{
FG_pull(Name)
}
\arguments{
\item{Name}{the name of the filegroup}
}
\value{
the filegroup dataframe object
}
\description{
Pull effort from a filegroup from the database by name. Uses INSTINCT format.
}
