% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{FGconvert}
\alias{FGconvert}
\title{Convert old effort data to filegroup}
\usage{
FGconvert(data, oldMooringName, source)
}
\arguments{
\item{data}{'Sfiles' type effort data as data frame}

\item{oldMooringName}{The mooring name of the old effort data.}

\item{source}{AFSC or DCLDE- different standards}
}
\value{
FG dataset
}
\description{
Using old (pre-NAS) effort data, convert to FG format.
}
