% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{GTconvert}
\alias{GTconvert}
\title{Convert old effort data to filegroup}
\usage{
GTconvert(
  GTdata,
  FGdat,
  Species,
  oldMooringName,
  labtype,
  vhz,
  a_id,
  source = "AFSC"
)
}
\arguments{
\item{GTdata}{The mooring name of the old effort data.}

\item{FGdat}{The data of type 'Sfile' corresponding to the GT data.}

\item{Species}{The two letter code corresponding to species/call type (RW, GS, etc)}

\item{labtype}{i_neg, det, etc.}

\item{vhz}{visible hz range of analysis (accepts 'max' as well as numeric)}

\item{a_id}{Integer ID for the analysis (must correspond to a record in 'analysis' table on the database)}

\item{source}{AFSC or DCLDE- different standards}
}
\value{
FG dataset
}
\description{
Using old (pre-NAS) effort data, convert to FG format.
}
