% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{RavenToDETx}
\alias{RavenToDETx}
\title{Convert a file from RAVENx to AFSC INSTINCT DETx format}
\usage{
RavenToDETx(FG, RAVx, out)
}
\arguments{
\item{FG}{A csv of AFSC INSTINCT datatype 'filegroup'.}

\item{RAVx}{A tab delimited Raven file containing the necessary RAVENx fields.}

\item{out}{full path of the file you want to save.}
}
\value{
DETx data object
}
\description{
Using a
}
